https://github.com/git-commit-id/maven-git-commit-id-plugin/issues/228: move first set of classes into new core module
