+++ core/src/main/java/pl/project13/core/git/GitDescribeConfig.java
@@ -15,7 +15,7 @@
  * along with git-commit-id-plugin.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package pl.project13.maven.git;
+package pl.project13.core.git;
 
 /**
  * Represents options passed in via maven configuration,
