+++ core/src/main/java/pl/project13/core/util/Pair.java
@@ -15,7 +15,7 @@
  * along with git-commit-id-plugin.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package pl.project13.maven.git.util;
+package pl.project13.core.util;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
