+++ maven/src/main/java/pl/project13/maven/git/GitDataProvider.java
@@ -18,10 +18,11 @@
 package pl.project13.maven.git;
 
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
-import pl.project13.maven.git.build.BuildServerDataProvider;
+import pl.project13.core.git.GitDescribeConfig;
 import pl.project13.maven.git.build.UnknownBuildServerData;
-import pl.project13.maven.git.log.LoggerBridge;
-import pl.project13.maven.git.util.PropertyManager;
+import pl.project13.maven.git.build.BuildServerDataProvider;
+import pl.project13.core.log.LoggerBridge;
+import pl.project13.core.util.PropertyManager;
 
 import javax.annotation.Nonnull;
 import java.net.URI;
