+++ maven/src/main/java/pl/project13/maven/git/PropertiesFilterer.java
@@ -20,7 +20,7 @@ package pl.project13.maven.git;
 import java.util.List;
 import java.util.Properties;
 
-import pl.project13.maven.git.log.LoggerBridge;
+import pl.project13.core.log.LoggerBridge;
 
 import javax.annotation.Nullable;
