+++ maven/src/test/java/pl/project13/core/jgit/DescribeCommandAbbrevIntegrationTest.java
@@ -15,14 +15,14 @@
  * along with git-commit-id-plugin.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package pl.project13.jgit;
+package pl.project13.core.jgit;
 
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.lib.Repository;
 import org.junit.Test;
+import pl.project13.core.log.StdOutLoggerBridge;
 import pl.project13.maven.git.AvailableGitTestRepo;
 import pl.project13.maven.git.GitIntegrationTest;
-import pl.project13.maven.git.log.StdOutLoggerBridge;
 
 import javax.annotation.Nonnull;
 import java.util.Optional;
