+++ maven/src/test/java/pl/project13/maven/git/GitPropertiesFileTest.java
@@ -27,7 +27,7 @@ import java.util.*;
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
 import org.junit.runner.RunWith;
-import pl.project13.maven.git.util.PropertyManager;
+import pl.project13.core.util.PropertyManager;
 
 import static java.util.Arrays.asList;
 import static org.fest.assertions.Assertions.assertThat;
