+++ maven/src/test/java/pl/project13/maven/git/NaivePerformanceTest.java
@@ -23,6 +23,7 @@ import org.apache.maven.project.MavenProject;
 import org.junit.Test;
 import org.junit.Ignore;
 import org.junit.runner.RunWith;
+import pl.project13.core.git.GitDescribeConfig;
 
 import java.util.Arrays;
 import java.util.Collection;
