+++ maven/src/test/java/pl/project13/maven/git/PropertiesFiltererTest.java
@@ -34,7 +34,7 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import pl.project13.maven.git.log.MavenLoggerBridge;
+import pl.project13.maven.log.MavenLoggerBridge;
 
 @RunWith(MockitoJUnitRunner.class)
 public class PropertiesFiltererTest {
